Add files via upload
The Project1.txt file showcases a slight graphic and textual form of what I wanted to code up, and the ScrollerRGB.java file holds the code for this. We can interact with sliders to change RGB values in a rectangle above the sliders, and pick a certain color that we need for an idea or a separate project. Also, there is a reset button implemented into this as well that will reset the sliders, and the rectangle back to 255 for all three values.